+++ controller/IPPPartner.php
@@ -173,7 +173,15 @@ class IPPPartner {
         return $this->request->curl($_ENV["GLOBAL_BASE_URL"]."/partner/communication/templates/standard/", "POST", [], $data);
     }
 
-
+    public function OnboardingData($company_id) {
+        $data = ["user_id" => $this->user_id, "session_id" => $this->session_id,"company_id" => $company_id];
+        return $this->request->curl($_ENV["GLOBAL_BASE_URL"]."/partner/company/data/onboarding/", "POST", [], $data)->content;
+    }
+    public function OnboardingPartnerData($company_id,$all_data,$application_state) {
+        $data = ["user_id" => $this->user_id, "session_id" => $this->session_id,"company_id"=>$company_id,"state" => $application_state];
+        $data = array_merge($all_data, $data);
+        return $this->request->curl($_ENV["GLOBAL_BASE_URL"]."/partner/company/data/onboarding/validation/", "POST", [], $data);
+    }
 
     public function RemovePlugin($id,$slug) {
         $data = ["user_id" => $this->user_id, "session_id" => $this->session_id,"plugin_id"=>$id,"plugin_slug"=>$slug];